spot_placement_score_lambda/requirements.txt
pyyaml>=5.0
boto3>=1.22.11
